core_requirements.txt
sqlalchemy==2.0.19
fastapi==0.100.0
uvicorn==0.23.1
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pre-commit==3.3.3
